refactor: relocate template catalog into its own agent module
