+++ apps/api/src/template-catalog/template-catalog.module.ts
@@ -1,9 +1,9 @@
 import { Module } from '@nestjs/common';
-import { WorkModule } from '@ever-works/agent/services';
+import { TemplateCatalogModule as AgentTemplateCatalogModule } from '@ever-works/agent/template-catalog';
 import { TemplateCatalogController } from './template-catalog.controller';
 
 @Module({
-    imports: [WorkModule],
+    imports: [AgentTemplateCatalogModule],
     controllers: [TemplateCatalogController],
 })
 export class TemplateCatalogModule {}
