+++ packages/agent/src/template-catalog/index.ts
@@ -1,0 +1,6 @@
+export { TemplateCatalogService } from './template-catalog.service';
+export type {
+    TemplateCatalogItem,
+    ForkTemplateResult,
+} from './template-catalog.service';
+export { TemplateCatalogModule } from './template-catalog.module';
